backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
apscheduler
pandas
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
